move test files to lib/
